Move FakeInfo to new file
