+++ src/PluginsSystem/Microsoft.Performance.Toolkit.Plugins.Runtime.Tests/TestData/FakeContents.cs
@@ -1,10 +1,7 @@
-﻿using Microsoft.Performance.Toolkit.Plugins.Core.Packaging.Metadata;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Microsoft.Performance.Toolkit.Plugins.Core;
+﻿// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+using Microsoft.Performance.Toolkit.Plugins.Core.Packaging.Metadata;
 
 namespace Microsoft.Performance.Toolkit.Plugins.Runtime.Tests
 {
@@ -15,20 +12,6 @@ namespace Microsoft.Performance.Toolkit.Plugins.Runtime.Tests
             return new PluginContents(
                 null,
                 null,
-                null);
-        }
-    }
-
-    internal static class FakeInfo
-    {
-        public static PluginInfo GetFakePluginInfoWithOnlyIdentityAndSdkVersion()
-        {
-            return new PluginInfo(
-                new PluginIdentity("fake_id", new Version("1.0.0")),
-                0,
-                null,
-                null,
-                new Version("1.0.0"),
                 null);
         }
     }
